% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_doc_count.R
\name{bind_doc_count}
\alias{bind_doc_count}
\title{Bind document count using inverse document frequency}
\usage{
bind_doc_count(tbl, document, idf)
}
\arguments{
\item{tbl}{A data frame, typically obtained from a call to \code{tidytext::bind_tf_idf()}.}

\item{document}{Column containing document identifiers, as a string or symbol.}

\item{idf}{Column containing the inverse document frequencies, as a string or symbol.}
}
\description{
Bind the count of the number of documents where a term occurs. The \code{tbl} argument is assumed to
be the output of a call to \code{tidytext::bind_tf_idf()}. This can be useful when filtering for noise
words that should be excluded from downstream analyses for not appearing in enough documents in a
corpus.
}
