% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonneg_stop_words.R
\docType{data}
\name{nonneg_stop_words}
\alias{nonneg_stop_words}
\title{Non-negative stop words}
\format{
A \link[tibble:tibble-package]{tibble} with 1094 rows and 2 columns:
\describe{
\item{word}{The non-negative stop word.}
\item{lexicon}{The lexicon of origin: SMART, onix, or snowball.}
}
}
\source{
Derived from \link[tidytext:stop_words]{tidytext::stop_words} with negation/negative terms removed.
}
\usage{
nonneg_stop_words
}
\description{
Stop word lexicons (SMART, onix, and snowball) but without negation words, i.e. negation words
will not be treated as stopwords. This is useful when looking for things such as non-compliance
in bigram analysis or topic modelling.
}
\keyword{datasets}
