% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-preview.R
\name{str_head}
\alias{str_head}
\alias{str_tail}
\title{Preview string beginnings and ends}
\usage{
str_head(string, width = NULL, ellipsis = "...")

str_tail(string, width = NULL, ellipsis = "...")
}
\arguments{
\item{string}{A character vector of strings to preview.}

\item{width}{A single integer for the number of characters to show, without consideration for the
length of \code{ellipsis}. If not supplied, defaults to the length of the longest string in \code{string}
divided by 10 and rounded up to the nearest integer.}

\item{ellipsis}{A single string used to indicate that content has been removed.}
}
\value{
A character vector the same length as \code{string}.
}
\description{
Use \code{str_head()} to view the beginning of a string and \code{str_tail()} to view the end of a string.
These functions work better for longer strings.
}
\examples{
# Basic usage
str_head(paste0(letters, collapse=""))
str_tail(paste0(letters, collapse=""))

# Slightly more realistic usage
library(tibble)
library(dplyr)
library(stringi)

my_docs <- tibble(
  doc_id = stri_rand_strings(4, length=6),
  text = rep(4, 4)
) \%>\%
mutate(text = stri_rand_lipsum(text, start_lipsum=FALSE))

my_docs \%>\%
  pull(text, name=doc_id) \%>\%
  str_head()

my_docs \%>\%
  pull(text, name=doc_id) \%>\%
  str_tail()
}
\seealso{
\code{\link[stringr:str_trunc]{stringr::str_trunc()}}, which this function wraps.
}
