% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringa-package.R
\docType{package}
\name{stringa-package}
\alias{stringa}
\alias{stringa-package}
\title{stringa: Adam's String Tools}
\description{
Tools that Adam often uses for his text-related projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adamoshen.github.io/stringa}
  \item \url{https://github.com/adamoshen/stringa}
}

}
\author{
\strong{Maintainer}: Adam Shen \email{adamshen1@gmail.com}

}
\keyword{internal}
