% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-preview.R
\name{str_head}
\alias{str_head}
\alias{str_tail}
\title{Preview string beginnings and ends}
\usage{
str_head(string, width = NULL, ellipsis = "...")

str_tail(string, width = NULL, ellipsis = "...")
}
\arguments{
\item{string}{A character vector of strings to preview.}

\item{width}{A single integer for the number of characters to show, without consideration for the
length of \code{ellipsis}. If not supplied, defaults to the length of the longest string in \code{string}
divided by 10 and rounded up to the nearest integer.}

\item{ellipsis}{A single string used to indicate that content has been removed.}
}
\value{
A character vector the same length as \code{string}.
}
\description{
Use \code{str_head()} to view the beginning of a string and \code{str_tail()} to view the end of a string.
These functions work better for longer strings.
}
\seealso{
\code{\link[stringr:str_trunc]{stringr::str_trunc()}}, which this function wraps.
}
